Fixing Clang tidy errors (#195)
* Fixing Clang tidy errors

* format-fix

* Update code_object.hpp

* Clang Tidy Fixes on the whole Source folder

* Update source/CMakeLists.txt


* Addressing reviews

* Correcting the logic for parsing att counters

* Format Fix

* Update source/lib/rocprofiler-sdk-att/tests/dummy_decoder.cpp


* Update source/lib/rocprofiler-sdk-att/tests/standalone_tool_main.cpp


* Update source/lib/rocprofiler-sdk-tool/config.cpp


* Formatting

* Deactivate clang-tidy in source/lib/rocprofiler-sdk-att/tests

---------


[ROCm/rocprofiler-sdk commit: 5410fabd3d502d17350548ef1f5eca9fc3b4f7ff]

+++ source/lib/rocprofiler-sdk-tool/tool.cpp
@@ -1394,7 +1394,8 @@ tool_init(rocprofiler_client_finalize_t fini_func, void* tool_data)
         parameters.push_back({ROCPROFILER_ATT_PARAMETER_SIMD_SELECT, {simd_select}});
         parameters.push_back({ROCPROFILER_ATT_PARAMETER_BUFFER_SIZE, {buffer_sz}});
         parameters.push_back({ROCPROFILER_ATT_PARAMETER_SHADER_ENGINE_MASK, {shader_mask}});
-        parameters.push_back({ROCPROFILER_ATT_PARAMETER_SERIALIZE_ALL, {att_serialize_all}});
+        parameters.push_back(
+            {ROCPROFILER_ATT_PARAMETER_SERIALIZE_ALL, {static_cast<uint64_t>(att_serialize_all)}});
 
         ROCPROFILER_CALL(
             rocprofiler_configure_dispatch_thread_trace_service(get_client_ctx(),
